Upgrade packages and manage lit import
