+++ package.json
@@ -16,7 +16,7 @@
   },
   "devDependencies": {
     "prettier": "^2.8.8",
-    "turbo": "^1.13.2"
+    "turbo": "^2.0.14"
   },
   "engines": {
     "node": ">=14.0.0"
