+++ packages/byfo-components-lit/package.json
@@ -4,7 +4,7 @@
   "version": "0.9.8",
   "type": "module",
   "scripts": {
-    "build": "tsc && vite build",
+    "build": "vite build",
     "tsc": "tsc",
     "preview": "vite preview",
     "generate": "node ./generate-element.js",
